+++ packages/tizen_bundle/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Increase the minimum Flutter version to 3.3.
+* Fix token in finalizer to be different from value.
 
 ## 0.1.1