+++ packages/tizen_bundle/CHANGELOG.md
@@ -1,7 +1,8 @@
-## NEXT
+## 0.1.2
 
 * Update minimum Flutter and Dart version to 3.13 and 3.1.
 * Fix token in finalizer to be different from value.
+* Update tizen_interop to 0.3.0.
 
 ## 0.1.1